+++ packages/amplify-cli-pm/CHANGELOG.md
@@ -1,3 +1,8 @@
+# v2.2.1
+
+ * chore: Updated AMPLIFY CLI references to Axway CLI.
+   ([CLI-100](https://jira.axway.com/browse/CLI-100))
+
 # v2.2.0 (Oct 1, 2020)
 
  * fix: Added back proxy server support.